+++ octasine/src/parameters/list.rs
@@ -96,6 +96,10 @@ pub const PARAMETERS: &[Parameter] = &[
     Parameter::Lfo(3, LfoParameter::Shape),
     Parameter::Lfo(3, LfoParameter::Amount),
     Parameter::Lfo(3, LfoParameter::Active),
+    Parameter::Lfo(0, LfoParameter::KeySync),
+    Parameter::Lfo(1, LfoParameter::KeySync),
+    Parameter::Lfo(2, LfoParameter::KeySync),
+    Parameter::Lfo(3, LfoParameter::KeySync),
 ];
 
 /// Parameter enum used to abstract over parameter indices
@@ -144,4 +148,6 @@ pub enum LfoParameter {
     Shape,
     Amount,
     Active,
+    /// Sync LFO phase to key presses. If turned off, start at random phase
+    KeySync,
 }